+++ resources/views/transfer/index.blade.php
@@ -70,9 +70,9 @@
                                 @endif
                                 <li class="divider"></li>
                                 @if(in_array("transfers-delete", $all_permission))
-                                {{ Form::open(['route' => ['transfers.destroy', $transfer->id], 'method' => 'DELETE'] ) }}
+                                {{ Form::open(['route' => ['transfers.destroy', $transfer->id], 'method' => 'DELETE','onsubmit' => 'return confirmDeleteAlert(this);'] ) }}
                                 <li>
-                                    <button type="submit" class="btn btn-link" onclick="return confirmDelete()"><i class="dripicons-trash"></i> {{trans('file.delete')}}</button>
+                                    <button type="submit" class="btn btn-link"  ><i class="dripicons-trash"></i> {{trans('file.delete')}}</button>
                                 </li>
                                 {{ Form::close() }}
                                 @endif
